+++ src/noctilux/backends/__init__.py
@@ -1,0 +1,3 @@
+from noctilux.backends.opencv_backend import cv2_to_pil, is_opencv_available, pil_to_cv2, require_opencv
+
+__all__ = ["is_opencv_available", "require_opencv", "pil_to_cv2", "cv2_to_pil"]